+++ client-app/src/Homepage/Convenors.jsx
@@ -1,0 +1,13 @@
+export const Convenors = () => {
+
+  return(
+      <div className='h-fit mb-10 bg-blue-500/10'>
+          <h1
+              className="p-4 md:p-10 pb-5 text-4xl md:text-5xl font-bold tracking-wider text-neutral-200 font-custom-sans"
+              data-aos="fade-up "
+          >
+              Convenors
+          </h1>
+      </div>
+  );
+}
